remove barrel exports in forms
