+++ packages/forms/src/components/BasicForm.stories.tsx
@@ -5,8 +5,8 @@ import { ReactNode } from "react";
 
 import BasicForm from "./BasicForm";
 import Repeater from "./Repeater/Repeater";
-import Submit from "./Submit";
-import Input from "./Input";
+import Submit from "./Submit/Submit";
+import Input from "./Input/Input";
 
 const labels = {
   requiredField: "Required field",
