--- packages/forms/src/components/CardOptionGroup/index.ts
@@ -1,5 +1,0 @@
-import CardOptionGroup from "./CardOptionGroup";
-import type { CardOptionGroupProps, CardOption } from "./CardOptionGroup";
-
-export default CardOptionGroup;
-export type { CardOptionGroupProps, CardOption };
