--- packages/forms/src/components/CheckButton/check-button.css
@@ -1,13 +1,0 @@
-.check-button {
-  border: none;
-  cursor: pointer;
-}
-
-.check-button__inner {
-  padding: 0.1rem;
-}
-
-.check-button__icon {
-  height: 0.75rem;
-  width: 0.75rem;
-}
