--- packages/forms/src/components/CheckButton/index.ts
@@ -1,4 +1,0 @@
-import CheckButton, { type CheckButtonProps } from "./CheckButton";
-
-export default CheckButton;
-export type { CheckButtonProps };
