--- packages/forms/src/components/Checkbox/index.ts
@@ -1,5 +1,0 @@
-import Checkbox from "./Checkbox";
-import type { CheckboxProps } from "./Checkbox";
-
-export default Checkbox;
-export type { CheckboxProps };
