+++ packages/forms/src/components/Checklist/Checklist.stories.tsx
@@ -4,7 +4,7 @@ import { action } from "@storybook/addon-actions";
 import { allModes } from "@gc-digital-talent/storybook-helpers";
 
 import Form from "../BasicForm";
-import Submit from "../Submit";
+import Submit from "../Submit/Submit";
 import Checklist from "./Checklist";
 
 export default {
