+++ packages/forms/src/components/Checklist/Checklist.tsx
@@ -2,7 +2,7 @@ import get from "lodash/get";
 import { FieldError, useFormContext } from "react-hook-form";
 import { ReactNode } from "react";
 
-import Checkbox from "../Checkbox";
+import Checkbox from "../Checkbox/Checkbox";
 import Field from "../Field";
 import type { CommonInputProps, HTMLFieldsetProps } from "../../types";
 import useFieldState from "../../hooks/useFieldState";
