--- packages/forms/src/components/Checklist/index.ts
@@ -1,5 +1,0 @@
-import Checklist from "./Checklist";
-import type { ChecklistProps, CheckboxOption } from "./Checklist";
-
-export default Checklist;
-export type { ChecklistProps, CheckboxOption };
