--- packages/forms/src/components/Combobox/index.ts
@@ -1,4 +1,0 @@
-import Combobox, { type ComboboxProps } from "./Combobox";
-
-export default Combobox;
-export type { ComboboxProps };
