+++ packages/forms/src/components/DateInput/DateInput.stories.tsx
@@ -19,7 +19,7 @@ import { fakePools } from "@gc-digital-talent/fake-data";
 import { allModes } from "@gc-digital-talent/storybook-helpers";
 
 import Form from "../BasicForm";
-import Submit from "../Submit";
+import Submit from "../Submit/Submit";
 import Input from "../Input/Input";
 import { DATE_SEGMENT } from "./types";
 import DateInput, { DateInputProps } from "./DateInput";
