--- packages/forms/src/components/Input/index.ts
@@ -1,5 +1,0 @@
-import Input from "./Input";
-import type { InputProps } from "./Input";
-
-export default Input;
-export type { InputProps };
