+++ packages/forms/src/components/RadioGroup/RadioGroup.stories.tsx
@@ -5,7 +5,7 @@ import { faker } from "@faker-js/faker/locale/en";
 import { VIEWPORT, allModes } from "@gc-digital-talent/storybook-helpers";
 
 import Form from "../BasicForm";
-import Submit from "../Submit";
+import Submit from "../Submit/Submit";
 import RadioGroup from "./RadioGroup";
 
 faker.seed(0);
