--- packages/forms/src/components/RadioGroup/index.ts
@@ -1,5 +1,0 @@
-import RadioGroup from "./RadioGroup";
-import type { RadioGroupProps, Radio } from "./RadioGroup";
-
-export default RadioGroup;
-export type { RadioGroupProps, Radio };
