+++ packages/forms/src/components/Repeater/Repeater.stories.tsx
@@ -9,8 +9,8 @@ import { errorMessages } from "@gc-digital-talent/i18n";
 import { allModes } from "@gc-digital-talent/storybook-helpers";
 
 import BasicForm from "../BasicForm";
-import Submit from "../Submit";
-import TextArea from "../TextArea";
+import Submit from "../Submit/Submit";
+import TextArea from "../TextArea/TextArea";
 import Repeater, { RepeaterFieldsetProps, RepeaterProps } from "./Repeater";
 
 type StoryProps = RepeaterProps &
