+++ packages/forms/src/components/RichTextInput/RichTextInput.stories.tsx
@@ -6,7 +6,7 @@ import { Heading } from "@gc-digital-talent/ui";
 import { allModes } from "@gc-digital-talent/storybook-helpers";
 
 import Form from "../BasicForm";
-import Submit from "../Submit";
+import Submit from "../Submit/Submit";
 import RichTextInput from "./RichTextInput";
 import RichTextRenderer from "./RichTextRenderer";
 import { htmlToRichTextJSON } from "../../utils";
