--- packages/forms/src/components/Select/index.ts
@@ -1,5 +1,0 @@
-import Select from "./Select";
-import type { SelectProps } from "./Select";
-
-export default Select;
-export type { SelectProps };
