--- packages/forms/src/components/Submit/index.ts
@@ -1,5 +1,0 @@
-import Submit from "./Submit";
-import type { SubmitProps } from "./Submit";
-
-export default Submit;
-export type { SubmitProps };
