+++ packages/forms/src/components/TextArea/TextArea.stories.tsx
@@ -4,10 +4,8 @@ import { StoryFn } from "@storybook/react";
 import { allModes } from "@gc-digital-talent/storybook-helpers";
 
 import Form from "../BasicForm";
-import Submit from "../Submit";
-
-import TextArea from ".";
-import type { TextAreaProps } from ".";
+import Submit from "../Submit/Submit";
+import TextArea, { type TextAreaProps } from "./TextArea";
 
 export default {
   component: TextArea,
