--- packages/forms/src/components/TextArea/index.ts
@@ -1,5 +1,0 @@
-import TextArea from "./TextArea";
-import type { TextAreaProps } from "./TextArea";
-
-export default TextArea;
-export type { TextAreaProps };
