--- packages/forms/src/components/WordCounter/index.ts
@@ -1,5 +1,0 @@
-import WordCounter from "./WordCounter";
-import type { WordCounterProps } from "./WordCounter";
-
-export default WordCounter;
-export type { WordCounterProps };
