+++ packages/forms/src/index.tsx
@@ -1,14 +1,16 @@
 import CardOptionGroup, {
   type CardOption,
   type CardOptionGroupProps,
-} from "./components/CardOptionGroup";
-import Checkbox, { type CheckboxProps } from "./components/Checkbox";
-import CheckButton, { CheckButtonProps } from "./components/CheckButton";
+} from "./components/CardOptionGroup/CardOptionGroup";
+import Checkbox, { type CheckboxProps } from "./components/Checkbox/Checkbox";
+import CheckButton, {
+  CheckButtonProps,
+} from "./components/CheckButton/CheckButton";
 import Checklist, {
   type ChecklistProps,
   type CheckboxOption,
-} from "./components/Checklist";
-import Combobox, { ComboboxProps } from "./components/Combobox";
+} from "./components/Checklist/Checklist";
+import Combobox, { ComboboxProps } from "./components/Combobox/Combobox";
 import DateInput, { DateInputProps } from "./components/DateInput/DateInput";
 import { DateSegment, DATE_SEGMENT } from "./components/DateInput/types";
 import Field, {
@@ -21,25 +23,24 @@ import Field, {
 import HiddenInput, {
   type HiddenInputProps,
 } from "./components/HiddenInput/HiddenInput";
-import Input, { type InputProps } from "./components/Input";
+import Input, { type InputProps } from "./components/Input/Input";
 import RadioGroup, {
   type RadioGroupProps,
   type Radio,
-} from "./components/RadioGroup";
+} from "./components/RadioGroup/RadioGroup";
 import Repeater from "./components/Repeater/Repeater";
 import RichTextInput from "./components/RichTextInput/RichTextInput";
 import RichTextRenderer from "./components/RichTextInput/RichTextRenderer";
-import Select, { type SelectProps } from "./components/Select";
-import Submit, { type SubmitProps } from "./components/Submit";
+import Select, { type SelectProps } from "./components/Select/Select";
+import Submit, { type SubmitProps } from "./components/Submit/Submit";
 import SwitchInput, {
   type SwitchInputProps,
 } from "./components/SwitchInput/SwitchInput";
-import TextArea, { type TextAreaProps } from "./components/TextArea";
-import WordCounter, { type WordCounterProps } from "./components/WordCounter";
-import BasicForm, {
-  type BasicFormProps,
-  type FieldLabels,
-} from "./components/BasicForm";
+import TextArea, { type TextAreaProps } from "./components/TextArea/TextArea";
+import WordCounter, {
+  type WordCounterProps,
+} from "./components/WordCounter/WordCounter";
+import BasicForm, { type BasicFormProps } from "./components/BasicForm";
 import ErrorSummary from "./components/ErrorSummary";
 import UnsavedChanges from "./components/UnsavedChanges";
 import {
@@ -57,7 +58,7 @@ import {
   alphaSortOptions,
 } from "./utils";
 import useInputStyles from "./hooks/useInputStyles";
-import { Option, OptGroup, OptGroupOrOption } from "./types";
+import { Option, OptGroup, OptGroupOrOption, FieldLabels } from "./types";
 
 export {
   DATE_SEGMENT,
