+++ packages/forms/src/types.ts
@@ -3,6 +3,7 @@ import {
   InputHTMLAttributes,
   FieldsetHTMLAttributes,
   ReactNode,
+  JSX,
 } from "react";
 import {
   RegisterOptions,
@@ -69,3 +70,29 @@ export type OptGroup = {
 };
 
 export type OptGroupOrOption = OptGroup | Option;
+
+export type FieldLabels = Record<string, ReactNode>;
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export interface Attrs {
+  readonly [attr: string]: any;
+}
+
+export interface Node {
+  type: string;
+  attrs?: Attrs;
+  marks?: Attrs[];
+  content?: Node[];
+  readonly [attr: string]: any;
+}
+
+export interface NodeProps {
+  children?: ReactNode;
+  node: Node;
+}
+
+export type NodeRenderer = (props: NodeProps) => JSX.Element;
+
+export interface RenderMap {
+  readonly [attr: string]: NodeRenderer;
+}
